Move tool registry plan tests into codex-tools
